+++ crates/client/src/commander/honggfuzz.rs
@@ -1,5 +1,4 @@
 use fehler::{throw, throws};
-use std::path::PathBuf;
 use std::process;
 use std::{os::unix::process::CommandExt, process::Stdio};
 use tokio::process::Command;
@@ -19,8 +18,6 @@ impl Commander {
         // obtain hfuzz_run_args from env variable, this variable can contain multiple
         // arguments so we need to parse the variable content.
         let hfuzz_run_args = std::env::var("HFUZZ_RUN_ARGS").unwrap_or_default();
-
-        let genesis_folder = PathBuf::from(self.root.to_string()).join("trident-genesis");
 
         let mut rustflags = std::env::var("RUSTFLAGS").unwrap_or_default();
 
@@ -50,7 +47,6 @@ impl Commander {
                 // enforce keep output to be true
                 fuzz_args.push_str("--keep_output");
                 let mut child = Command::new("cargo")
-                    .env("GENESIS_FOLDER", genesis_folder)
                     .env("HFUZZ_RUN_ARGS", fuzz_args)
                     .env("CARGO_TARGET_DIR", cargo_target_dir)
                     .env("HFUZZ_WORKSPACE", hfuzz_workspace)
@@ -64,7 +60,6 @@ impl Commander {
             }
             false => {
                 let mut child = Command::new("cargo")
-                    .env("GENESIS_FOLDER", genesis_folder)
                     .env("HFUZZ_RUN_ARGS", fuzz_args)
                     .env("CARGO_TARGET_DIR", cargo_target_dir)
                     .env("HFUZZ_WORKSPACE", hfuzz_workspace)
@@ -95,8 +90,6 @@ impl Commander {
 
         let hfuzz_run_args = std::env::var("HFUZZ_RUN_ARGS").unwrap_or_default();
 
-        let genesis_folder = PathBuf::from(self.root.to_string()).join("trident-genesis");
-
         let cargo_target_dir = std::env::var("CARGO_TARGET_DIR")
             .unwrap_or_else(|_| config.get_env_arg(&EnvVariable::CargoTargetDir));
         let hfuzz_workspace = std::env::var("HFUZZ_WORKSPACE")
@@ -113,7 +106,6 @@ impl Commander {
                 // enforce keep output to be true
                 fuzz_args.push_str("--keep_output");
                 let mut child = Command::new("cargo")
-                    .env("GENESIS_FOLDER", genesis_folder)
                     .env("HFUZZ_RUN_ARGS", fuzz_args)
                     .env("CARGO_TARGET_DIR", cargo_target_dir)
                     .env("HFUZZ_WORKSPACE", hfuzz_workspace)
@@ -127,7 +119,6 @@ impl Commander {
             }
             false => {
                 let mut child = Command::new("cargo")
-                    .env("GENESIS_FOLDER", genesis_folder)
                     .env("HFUZZ_RUN_ARGS", fuzz_args)
                     .env("CARGO_TARGET_DIR", cargo_target_dir)
                     .env("HFUZZ_WORKSPACE", hfuzz_workspace)
@@ -147,8 +138,6 @@ impl Commander {
 
         let crash_file = std::path::Path::new(&self.root as &str).join(crash_file_path);
 
-        let genesis_folder = PathBuf::from(self.root.to_string()).join("trident-genesis");
-
         if !crash_file.try_exists()? {
             println!("{ERROR} The crash file [{:?}] not found", crash_file);
             throw!(Error::CrashFileNotFound);
@@ -163,7 +152,6 @@ impl Commander {
 
         // using exec rather than spawn and replacing current process to avoid unflushed terminal output after ctrl+c signal
         std::process::Command::new("cargo")
-            .env("GENESIS_FOLDER", genesis_folder)
             .env("CARGO_TARGET_DIR", cargo_target_dir)
             .env("RUSTFLAGS", rustflags)
             .arg("hfuzz")