+++ examples/fuzz-tests/arbitrary-custom-types-4/trident-tests/fuzz_tests/fuzz_0/test_fuzz.rs
@@ -1,7 +1,6 @@
 use fuzz_instructions::Initialize;
 use fuzz_instructions::Update;
 use trident_client::fuzzing::*;
-
 mod fuzz_instructions;
 
 use arbitrary_custom_types_4::entry as entry_arbitrary_custom_types_4;
@@ -31,14 +30,16 @@ fn fuzz_iteration<T: FuzzTestExecutor<U> + std::fmt::Display, U>(
     );
 
     let mut client =
-        ProgramTestClientBlocking::new(&[fuzzing_program_arbitrary_custom_types_4], &[]).unwrap();
+        ProgramTestClientBlocking::new(&[fuzzing_program_arbitrary_custom_types_4], config)
+            .unwrap();
 
     let _ = fuzz_data.run_with_runtime(&mut client, config);
 }
 
 fn main() {
     let config = Config::new();
+
     loop {
-        fuzz_trident ! (fuzz_ix : FuzzInstruction , | fuzz_data : MyFuzzData | { fuzz_iteration (fuzz_data,&config) ; });
+        fuzz_trident ! (fuzz_ix : FuzzInstruction , | fuzz_data : MyFuzzData | { fuzz_iteration (fuzz_data , & config) ; });
     }
 }