+++ examples/fuzz-tests/unauthorized-access-2/trident-tests/fuzz_tests/fuzz_0/test_fuzz.rs
@@ -30,14 +30,15 @@ fn fuzz_iteration<T: FuzzTestExecutor<U> + std::fmt::Display, U>(
     );
 
     let mut client =
-        ProgramTestClientBlocking::new(&[fuzzing_program_unauthorized_access_2], &[]).unwrap();
+        ProgramTestClientBlocking::new(&[fuzzing_program_unauthorized_access_2], config).unwrap();
 
     let _ = fuzz_data.run_with_runtime(&mut client, config);
 }
 
 fn main() {
     let config = Config::new();
+
     loop {
-        fuzz_trident ! (fuzz_ix : FuzzInstruction , | fuzz_data : MyFuzzData | { fuzz_iteration (fuzz_data,&config) ; });
+        fuzz_trident ! (fuzz_ix : FuzzInstruction , | fuzz_data : MyFuzzData | { fuzz_iteration (fuzz_data , & config) ; });
     }
 }